pantryai/requirements.txt
flask==3.0.3
flask-cors==4.0.1
boto3==1.34.100
requests==2.31.0
opensearch-py==2.5.0
python-dotenv==1.0.1
gunicorn==22.0.0
requests-aws4auth==1.2.3
